requirements.txt
django==2.0.8
djangorestframework
markdown
django-filter
django-cors-headers
